+++ comics/comics/havet.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import CrawlerBase
 from comics.core.comic_data import ComicDataBase
 
 
@@ -9,15 +9,11 @@ class ComicData(ComicDataBase):
     language = 'no'
     url = 'http://havet.nettserier.no/'
     start_date = '2007-09-27'
+    end_date = '2012-10-25'
+    active = False
     rights = 'Øystein Ottesen'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 90
-    schedule = 'We'
-    time_zone = 'Europe/Oslo'
-
     def crawl(self, pub_date):
-        epoch = self.date_to_epoch(pub_date)
-        url = 'http://havet.nettserier.no/_striper/havet-%s.jpg' % epoch
-        return CrawlerImage(url)
+        pass  # Comic no longer published